% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_percent_format}
\alias{calc_percent_format}
\title{Add percentages to a table}
\usage{
calc_percent_format(input.tbl)
}
\arguments{
\item{input.tbl}{Table to be modified}
}
\value{
A table with the original cell-wise totals and percentages by row in parentheses
}
\description{
Add percentages to a table
}
