% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_pval_dem}
\alias{format_pval_dem}
\title{Round p-values and indicate significance with asterisks}
\usage{
format_pval_dem(p.val)
}
\arguments{
\item{p.val}{P-value to be modified}
}
\value{
Formatted p-value, rounded to the second decimal place and with asterisks to indicate significance
}
\description{
Round p-values and indicate significance with asterisks
}
