% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_summary_stats}
\alias{format_summary_stats}
\title{Summarize a vector by mean, standard deviation, median, and range}
\usage{
format_summary_stats(input.vector)
}
\arguments{
\item{input.vector}{Numeric vector to be sumarrized}
}
\value{
Vector of summary statistics
}
\description{
Summarize a vector by mean, standard deviation, median, and range
}
