% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_table.R
\name{strat_dem_tbl}
\alias{strat_dem_tbl}
\title{Create a stratified demographics table}
\usage{
strat_dem_tbl(data, outcome, continuous.covariates = NULL,
  continuous.tests = "wilcox.test", categorical.covariates = NULL,
  categorical.tests = "likelihood.ratio.chi.square", strata.col = NULL)
}
\arguments{
\item{data}{A set of data with all covariates in separate columns and each row representing exactly 1 individual}

\item{outcome}{Column containing indicating the primary outcome of interest (e.g. case/control or disease subtypes)}

\item{continuous.covariates}{Columns containing continuous covariates of interest}

\item{continuous.tests}{A vector of tests to run on the continuous covariates}

\item{categorical.covariates}{Columns containing the categorical covariates of interest}

\item{categorical.tests}{A vector of tests to run on categorical covariates}

\item{strata.col}{Column containing the categories by which the data will be stratified}
}
\value{
A demographics table in the format typically implemented by the iMAGE study.
}
\description{
Create a stratified demographics table
}
