% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{count_NAs_in_tbl}
\alias{count_NAs_in_tbl}
\title{Count the NA values in a table}
\usage{
count_NAs_in_tbl(input.tbl)
}
\arguments{
\item{input.tbl}{Table in which to count NA values}
}
\value{
A vector, the first value of which is the total number of NAs for all
outcome categories. The remaining values are the category-specific number of
missing values.
}
\description{
Count the NA values in a table
}
